feat(Client): new Solution for reload weak labelling and fetch rule only in tokenClassification task
